pyproject.toml
[project]
name = "compliance-leaderboard"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "httpx>=0.28.0",
    "instructor>=1.7.0",
    "litellm>=1.56.0",
    "openai>=1.60.0",
    "pandas>=2.2.0",
    "pydantic>=2.10.0",
    "pymupdf4llm>=0.0.17",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.6.0",
    "streamlit>=1.41.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.8.0",
    "tqdm>=4.67.0",
]

[tool.uv]
package = false
